clear order_id session (mirror CheckoutController)
I noticed a lack of symmetry between the update method in Spree::CheckoutController and here. 
https://github.com/spree/spree/pull/2557

More generally the problem is that we have duplication of the success code within Paypal Express and within Spree Core code. This has caused problems in the past. A possible refactor would be to slap the four lines of completion code (e.g. setting the flash ecommerce) into a Spree Core module (exiting or new) that gets pulled into this and checkout controller.