Use discounted shipping cost instead of ship_cost.
PayPal will reject orders which do not have their totals line up.  If a
Free Shipping promotion is applied to an order, the initial shipping
cost will be used and the free shipping promotion will not be accounted
for.

This will cause the order to get rejected by PayPal because the order
total doesn't match the sum of the line items.

Discounted ship cost should be the correct value to use here.

Fixes #99