Only ignore shipping promotions.
The code was working on the fact that shipping adjustments get rolled into the
shipping line item, however the case is that code uses `discount_total`, which
is `cost` + `promo_cost` (which is best_promotion_adjustment).

Since we're adding premium packaging as a shipment adjustment (that's not a
promotion), we really shouldn't be ignoring it.